+++ src/main/java/com/codeit/sb01otbooteam06/domain/clothes/service/ClothesService.java
@@ -27,8 +27,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -41,6 +39,7 @@ import org.springframework.web.multipart.MultipartFile;
 public class ClothesService {
 
   private final S3Service s3Service;
+  private final ClothesCacheService clothesCacheService;
 
   private final ClothesRepository clothesRepository;
   private final UserRepository userRepository;
@@ -89,7 +88,8 @@ public class ClothesService {
 
     List<ClothesAttribute> clothesAttributes = clothesAttributeService.create(clothes,
         clothesCreateRequest.attributes());
-    invalidateUserClothesCache(owner.getId(
+
+    clothesCacheService.invalidateUserClothesCache(owner.getId(
     )); //의상 캐시 삭제
 
     return customClothesUtils.makeClothesDto(clothes, clothesAttributes);
@@ -213,7 +213,7 @@ public class ClothesService {
         .orElseThrow(() -> new ClothesNotFoundException().withId(clothesId));
     UUID userId = clothesRepository.findById(clothesId).get().getOwner().getId();
     clothesRepository.deleteById(clothesId);
-    invalidateUserClothesCache(userId); //의상 개수 캐시 삭제
+    clothesCacheService.invalidateUserClothesCache(userId); //의상 개수 캐시 삭제
 
   }
 
@@ -262,22 +262,6 @@ public class ClothesService {
 
     return new ClothesDto(null, null, null, null, null, null);
 
-  }
-
-
-  //유저 옷장의 옷 개수 캐시 //todo 페이지네이션에서 이용하기!
-  @Cacheable(value = "userClothesCount", key = "#userId")
-  public int getUserClothesCountWithCache(UUID userId) {
-    return clothesRepository.getTotalCounts("", userId);
-  }
-
-  public int getUserClothesCount(UUID userId) {
-    return clothesRepository.getTotalCounts("", userId);
-  }
-
-  // 호출시 유저의 의상 수 캐시 제거됨 (예: 옷 추가/삭제 시)
-  @CacheEvict(value = "userClothesCount", key = "#userId")
-  public void invalidateUserClothesCache(UUID userId) {
   }
